pyproject.toml
[tool.poetry]
name = "xenia-admin"
version = "0.1.0"
description = ""
authors = ["RobertoChiosa <roberto@xeniapm.it>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
django-timescaledb = "^0.2.13"
django-simple-history = "^3.7.0"
django-import-export = "^4.0.9"
django = "^5.1.2"
psycopg2-binary = "^2.9.9"
pychartjs = "^0.9.2"
pandas = "^2.2.2"
defusedxml = "^0.7.1"
et-xmlfile = "^1.1.0"
markuppy = "^1.14"
openpyxl = "^3.1.4"
django-unfold = "^0.40.0"
import-export = "^0.3.1"
whitenoise = "^6.7.0"
django-modeltranslation = "^0.19.11"
django-sass-processor = "^1.4.1"
reportlab = "^4.2.5"
fpdf2 = "^2.8.1"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
import_heading_stdlib = "Standard library imports"
import_heading_thirdparty = "Third party imports"
import_heading_firstparty = "Project imports"

